gui/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

PyQt5==5.15.10
pyserial==3.5
